Use perf_counter for measurements; running avg for live timing display; move display helpers to display_utils
- multiturn_env, rubric, rubric_group, env_group: switch all timing
  measurements from time.time() to time.perf_counter() (monotonic). Use
  the start_time/end_time pattern with end-time captured before writing
  into state["timing"], and += for running totals.
- eval_utils.on_display_progress: keep running timing_sums/timing_counts
  in the closure and update from new_outputs only — was O(N) per
  progress event, now O(M) per event / O(N) over the run.
- Drop "(avg)" suffix on the non-TUI summary "Timing:" line.
- Move _timing_parts / format_timing_line / format_timing_rich from
  logging_utils.py to display_utils.py.
