+++ verifiers/utils/eval_utils.py
@@ -637,7 +637,7 @@ def print_info(results: GenerateOutputs):
 
 
 def print_timing(results: GenerateOutputs):
-    from verifiers.utils.logging_utils import format_timing_line
+    from verifiers.utils.display_utils import format_timing_line
 
     outputs = results["outputs"]
     timing_list = [o["timing"] for o in outputs]
@@ -655,7 +655,7 @@ def print_timing(results: GenerateOutputs):
     avg_env = float(np.mean(env_s_totals)) if env_s_totals else None
 
     print(
-        "Timing (avg): "
+        "Timing: "
         + format_timing_line(
             total=_mean("total"),
             setup=_mean("setup"),
@@ -976,6 +976,20 @@ async def run_evaluations_tui(
                 env_idx, total=total, num_examples=num_examples, progress=resumed
             )
 
+        # Running sums for live timing average — only walk new outputs each
+        # progress event (O(M) per update, O(N) total across the run).
+        timing_sums: dict[str, float] = {}
+        timing_counts: dict[str, int] = {}
+        TIMING_KEYS = (
+            "setup",
+            "generation",
+            "scoring",
+            "overhead",
+            "total",
+            "model",
+            "env",
+        )
+
         def on_display_progress(
             all_outputs: list[RolloutOutput],
             new_outputs: list[RolloutOutput],
@@ -989,31 +1003,19 @@ async def run_evaluations_tui(
             for k, v in pass_all_k.items():
                 metrics[f"pass^{k}"] = v
 
-            # Compute average timing across all outputs
-            avg_timing = None
-            if all_outputs:
-                timing_sums: dict[str, float] = {}
-                timing_counts: dict[str, int] = {}
-                for o in all_outputs:
-                    t = o.get("timing")
-                    if not isinstance(t, dict):
-                        continue
-                    for key in (
-                        "setup",
-                        "generation",
-                        "scoring",
-                        "overhead",
-                        "total",
-                        "model",
-                        "env",
-                    ):
-                        if key in t:
-                            timing_sums[key] = timing_sums.get(key, 0.0) + float(t[key])
-                            timing_counts[key] = timing_counts.get(key, 0) + 1
-                if timing_sums:
-                    avg_timing = {
-                        k: timing_sums[k] / timing_counts[k] for k in timing_sums
-                    }
+            for o in new_outputs:
+                t = o.get("timing")
+                if not isinstance(t, dict):
+                    continue
+                for key in TIMING_KEYS:
+                    if key in t:
+                        timing_sums[key] = timing_sums.get(key, 0.0) + float(t[key])
+                        timing_counts[key] = timing_counts.get(key, 0) + 1
+            avg_timing = (
+                {k: timing_sums[k] / timing_counts[k] for k in timing_sums}
+                if timing_sums
+                else None
+            )
 
             display.update_env_state(
                 env_idx,
